backend/requirements.txt
# Flask Server Dependencies
flask
flask-cors
python-dotenv
groq
